after migration fixes

+++ Application/DTOs/UserSide/ShowAllDto.cs
@@ -14,7 +14,7 @@ namespace Application.DTOs.UserSide
         public string TitleDescription { get; set; }
         public string Description { get; set; }
         public string Email { get; set; }
-        public int MobilePhone { get; set; }
+        public long MobilePhone { get; set; }
         public string Location { get; set; }
         public string PicName { get; set; }
         #endregion
